requirements.txt
docxtpl==0.20.1
pyutils @ git+https://github.com/iagocanalejas/pyutils.git@master
pyyaml==6.0.3
